backend/requirements.txt
google-adk==1.0.0
fastapi
uvicorn[standard]
python-dotenv
pytest
pytest-asyncio
httpx
websockets
chromadb
